Rename driver/analysis.py to _analysis.py to match develop
The develop branch renamed the driver modules with an underscore prefix
(analysis.py -> _analysis.py). Making the rename explicit here aligns the
file path on both sides so the PR merges without a rename/modify conflict.
Update the driver __init__ import to match.
